refactor: share vue plugins
