+++ packages/vue-language-core/src/index.ts
@@ -11,3 +11,7 @@ export * from './parsers/scriptSetupRanges';
 export * as localTypes from './utils/localTypes';
 export * from './utils/ts';
 export * from './generators/template';
+
+// fix build
+export * as _0 from '@vue/compiler-sfc';
+export * as _1 from '@vue/compiler-core';
