+++ packages/vue-language-service/src/documentService.ts
@@ -104,6 +104,7 @@ export function getDocumentService(
 			}
 		},
 	};
+	const vuePlugins = vue.getPlugins(ts, {}, {}, []);
 
 	return {
 		format: format.register(context),
@@ -131,8 +132,8 @@ export function getDocumentService(
 			'/untitled.' + shared.languageIdToSyntax(document.languageId),
 			document.getText(),
 			{},
-			{},
 			context.typescript,
+			vuePlugins,
 		);
 		vueDoc = parseVueDocument(vueFile, undefined);
